// packages/core/src/types/consent-types.d.ts
export type AllConsentNames =
	| 'experience'
	| 'functionality'
	| 'marketing'
	| 'measurement'
	| 'necessary';
// ...
export declare const allConsentNames: AllConsentNames[];